Added two common component
