+++ src/components/teachers/Teachers.js
@@ -62,28 +62,6 @@ const useStyles = makeStyles((theme) => ({
       minWidth: '260px'
     }
   },
-  titlePage: {
-    color: 'black',
-    fontWeight: '500',
-    fontFamily: 'Roboto',
-    width: '92.5%',
-    [theme.breakpoints.down('xs')]: {
-      width: '59%',
-    }
-  },
-  titlePageHide: {
-    width: '93.4%',
-    [theme.breakpoints.down('xs')]: {
-      width: '70%',
-    }
-  },
-  header: {
-    display: 'flex'
-  },
-  modal: {
-    paddingTop: theme.spacing(2.2),
-    width: '100px'
-  }
 }))
 
 const Teachers = ({show}) => {
@@ -108,19 +86,14 @@ const Teachers = ({show}) => {
   const classes = useStyles()
   return (
     <Box className={show ? classes.table : classes.tableHide}>
-      <Box className={classes.header}>
-        <h2 className={show ? classes.titlePage : classes.titlePageHide}>Teachers</h2>
-        <Box className={classes.modal}>
-          <TeachersModal 
-            open = {open}
-            setOpen = {setOpen}
-            loadData = {loadData} 
-            editForm = {editForm}
-            handleCloseModal = {handleCloseModal}  
-            setEditForm = {setEditForm} 
-          />  
-        </Box>
-      </Box>
+      <TeachersModal 
+        open = {open}
+        setOpen = {setOpen}
+        loadData = {loadData} 
+        editForm = {editForm}
+        handleCloseModal = {handleCloseModal}  
+        setEditForm = {setEditForm} 
+      />  
       <TableTeachers 
         teachers = {teachers}
         loadData = {loadData}
