+++ src/components/teachers/TeachersModal.js
@@ -1,21 +1,9 @@
-import Button from '@mui/material/Button';
 import { styled } from '@mui/material/styles';
 import Dialog from '@mui/material/Dialog';
 import DialogContent from '@mui/material/DialogContent';
 import { Box } from '@mui/material';
 import ManageTeacherForm from '../teachers/ManageTeacherForm';
-import { makeStyles } from '@material-ui/core'
-
-const useStyles = makeStyles(() => ({
-  button: {
-    width: '110px',
-    height: '40px',
-    color: 'black',
-  },
-  buttonName: {
-    color: 'black',
-  }
-}))
+import HeaderCommon from '../common/HeaderCommon';
 
 const BootstrapDialog = styled(Dialog)(({ theme }) => ({
   '& .MuiDialogContent-root': {
@@ -27,18 +15,15 @@ const BootstrapDialog = styled(Dialog)(({ theme }) => ({
 }));
 
 export default function CustomizedDialogs({open, setOpen, courses, loadData, editForm, handleCloseModal, setEditForm}) {
-  const classes = useStyles()
   return (
-    <Box className='modal'>
-      <Button  
+    <Box>
+      <HeaderCommon  
         onClick={() => {
           setOpen(true)
-          setEditForm({})}}
-          className={classes.button}
-          variant="outlined"
-          >
-          <p className={classes.buttonName}>Add New</p>
-      </Button>
+          setEditForm({})}
+        }
+        text = "Teachers"
+      />
       <BootstrapDialog
         onClose={() => setOpen(false)}
         aria-labelledby="customized-dialog-title"
